HCS Topic Listener (#438)
* HCS Topic Listener utilizing pg notify flux


* Added tests for message filter. Took Stevens improved builder logic and cleaned up according to feedback.


* Fixes


* Missing semicolon


* More tests and fixes


* Touched up test cases and addressed filter case where time is on border


* Query for missing messages


* Message order fixes


* Make tests more deterministic


* Fix issue when filter.endTime before start

+++ hedera-mirror-grpc/src/main/java/com/hedera/mirror/grpc/domain/TopicMessageFilter.java
@@ -41,14 +41,14 @@ public class TopicMessageFilter {
     private long limit;
 
     @Min(0)
-    private long realmNum;
+    private int realmNum;
 
     @NotNull
     @Builder.Default
     private Instant startTime = Instant.now();
 
     @Min(0)
-    private long topicNum;
+    private int topicNum;
 
     public boolean hasLimit() {
         return limit > 0;
